hafta_4/requirements.txt
sentence-transformers==2.2.2
scikit-learn==1.3.0
matplotlib==3.7.2
numpy==1.24.3
faiss-cpu==1.7.4
chromadb==0.4.15
psutil==5.9.5
openai==1.3.0
python-dotenv==1.0.0